+++ OOP/Lesson_5.h
@@ -1,0 +1,5 @@
+#pragma once
+void Lesson_5()
+{
+}
+
